:domain 경로 변경
