+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/constant/OAuthType.kt
@@ -1,0 +1,6 @@
+package com.example.sachosaeng.core.domain.constant
+
+enum class OAuthType {
+    GOOGLE,
+    KAKAO
+}
