+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/model/Category.kt
@@ -1,0 +1,6 @@
+package com.example.sachosaeng.core.domain.model
+data class Category(
+    val id: Int = 0,
+    val name: String = "카테고리",
+    val imageUrl: String? = null,
+)
