+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/model/User.kt
@@ -1,0 +1,10 @@
+package com.example.sachosaeng.core.domain.model
+
+import com.example.sachosaeng.core.domain.constant.OAuthType
+
+data class User(
+    val id: Int,
+    val name: String,
+    val email: String,
+    val oAuthType: OAuthType
+)
