+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/repository/CardRepository.kt
@@ -1,0 +1,10 @@
+package com.example.sachosaeng.core.domain.repository
+
+import com.example.sachosaeng.core.domain.model.Card
+
+
+interface CardRepository {
+    suspend fun getCardList(): List<Card>
+    suspend fun getMyCardList(): List<Card>
+    suspend fun bookmarkCard(card: Card)
+}
