+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/repository/RequestToAdminRepository.kt
@@ -1,0 +1,5 @@
+package com.example.sachosaeng.core.domain.repository
+
+interface RequestToAdminRepository {
+    suspend fun sendRequestToAdmin(request: String);
+}
