+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/repository/SettingRepository.kt
@@ -1,0 +1,7 @@
+package com.example.sachosaeng.core.domain.repository
+
+interface SettingRepository {
+    suspend fun getPrivacyPolicyUrl(): String
+    suspend fun getTermsOfServiceUrl(): String
+    suspend fun updateVersion()
+}
