+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/usecase/Usecase.kt
@@ -1,0 +1,7 @@
+package com.example.sachosaeng.core.domain.usecase
+
+import java.util.Optional
+
+interface Usecase<RESULT> {
+    operator suspend fun invoke(): RESULT
+}
