+++ core/domain/src/main/java/com/example/sachosaeng/core/domain/usecase/user/WithdrawUsecase.kt
@@ -1,0 +1,8 @@
+package com.example.sachosaeng.core.domain.usecase.user
+
+import com.example.sachosaeng.core.domain.repository.UserRepository
+import com.example.sachosaeng.core.domain.usecase.Usecase
+
+class WithdrawUsecase(private val repositroy : UserRepository) : Usecase<Unit> {
+    override suspend fun invoke() = repositroy.withdraw()
+}
