--- core/domain/src/main/java/com/example/sachosaeng/domain/constant/OAuthType.kt
@@ -1,6 +1,0 @@
-package com.example.sachosaeng.domain.constant
-
-enum class OAuthType {
-    GOOGLE,
-    KAKAO
-}
