--- core/domain/src/main/java/com/example/sachosaeng/domain/model/Card.kt
@@ -1,7 +1,0 @@
-data class Card(
-    val id: Int,
-    val title: String,
-    val category: Category,
-    val content: String,
-    val isBookmarked: Boolean,
-)
