--- core/domain/src/main/java/com/example/sachosaeng/domain/model/Category.kt
@@ -1,4 +1,0 @@
-data class Category(
-    val id: Int,
-    val name: String,
-)
