--- core/domain/src/main/java/com/example/sachosaeng/domain/model/User.kt
@@ -1,8 +1,0 @@
-import com.example.sachosaeng.domain.constant.OAuthType
-
-data class User(
-    val id: Int,
-    val name: String,
-    val email: String,
-    val oAuthType: com.example.sachosaeng.domain.constant.OAuthType
-)
