--- core/domain/src/main/java/com/example/sachosaeng/domain/repository/CardRepository.kt
@@ -1,9 +1,0 @@
-package com.example.sachosaeng.domain.repository
-
-import Card
-
-interface CardRepository {
-    suspend fun getCardList(): List<Card>
-    suspend fun getMyCardList(): List<Card>
-    suspend fun bookmarkCard(card: Card)
-}
