--- core/domain/src/main/java/com/example/sachosaeng/domain/repository/CategoryRepository.kt
@@ -1,9 +1,0 @@
-package com.example.sachosaeng.domain.repository
-
-import Category
-
-interface CategoryRepository {
-    suspend fun getCategoryList(): List<Category>
-    suspend fun getMyCategoryList(): List<Category>
-    suspend fun setCategoryList(categoryList: List<Category>)
-}
