--- core/domain/src/main/java/com/example/sachosaeng/domain/repository/RequestToAdminRepository.kt
@@ -1,7 +1,0 @@
-package com.example.sachosaeng.domain.repository
-
-import Card
-
-interface RequestToAdminRepository {
-    suspend fun sendRequestToAdmin(request: String);
-}
