--- core/domain/src/main/java/com/example/sachosaeng/domain/repository/SettingRepository.kt
@@ -1,7 +1,0 @@
-package com.example.sachosaeng.domain.repository
-
-interface SettingRepository {
-    suspend fun getPrivacyPolicyUrl(): String
-    suspend fun getTermsOfServiceUrl(): String
-    suspend fun updateVersion()
-}
