--- core/domain/src/main/java/com/example/sachosaeng/domain/repository/UserRepository.kt
@@ -1,12 +1,0 @@
-package com.example.sachosaeng.domain.repository
-
-import User
-
-interface UserRepository {
-    suspend fun getMyInfo(): User
-    fun login(id: String, password: String)
-    fun logout()
-    fun signIn(user: User)
-    fun updateMyInfo(user: User)
-    suspend fun withdraw()
-}
