--- core/domain/src/main/java/com/example/sachosaeng/domain/usecase/Usecase.kt
@@ -1,7 +1,0 @@
-package com.example.sachosaeng.domain.usecase
-
-import java.util.Optional
-
-interface Usecase<RESULT> {
-    operator suspend fun invoke(): RESULT
-}
