--- core/domain/src/main/java/com/example/sachosaeng/domain/usecase/user/LoginUsecase.kt
@@ -1,9 +1,0 @@
-package com.example.sachosaeng.domain.usecase.user
-
-import com.example.sachosaeng.domain.repository.UserRepository
-import com.example.sachosaeng.domain.usecase.Usecase
-
-class LoginUsecase(private val repository: com.example.sachosaeng.domain.repository.UserRepository, val id: String, val password: String) :
-    com.example.sachosaeng.domain.usecase.Usecase<Unit> {
-    override suspend fun invoke() = repository.login(id = id, password = password)
-}
