--- core/domain/src/main/java/com/example/sachosaeng/domain/usecase/user/WithdrawUsecase.kt
@@ -1,9 +1,0 @@
-package com.example.sachosaeng.domain.usecase.user
-
-import com.example.sachosaeng.domain.repository.UserRepository
-import com.example.sachosaeng.domain.usecase.Usecase
-
-class WithdrawUsecase(private val repositroy : com.example.sachosaeng.domain.repository.UserRepository) :
-    com.example.sachosaeng.domain.usecase.Usecase<Unit> {
-    override suspend fun invoke() = repositroy.withdraw()
-}
