+++ feature/build.gradle.kts
@@ -96,4 +96,5 @@ dependencies {
 
     implementation(project(":core:ui"))
     implementation(project(":core:util"))
+    implementation(project(":core:domain"))
 }
